Very big refactoring of the led_strip and driver settings
CAUTION: THIS VERSION IS NOT COMPATIBLE WITH PREVIOUS VERSIONS

To adapt to this version only rather small changes are required The biggest change is to not use atoms for drivers, but modules and it defines the default settings. Those can be overwritten through keywords.

This commit does NOT fix all remaining issues after this refactoring, but it's functional enough. Check the TODOs for the remaining parts