This is a breaking changes (even though only small adjustments are required to fix them.
Removed the (in the meantime, due to the DSL) unnecessary checks in the led_strip.

started to harmonize the interface to the Leds functions. They now take the optional parameters as opts instead of having functions with different arities. (Note: the rainbow function will come as next, that one is not done yet)

Started to overhaul the livebooks, but that will take some more iterations. Also they are not fully tested (an extra round will be required)