Introducing the Fledex.Color protocol
This will remove the issues with cyclic dependencies, allows to remove some duplication but does mean a bit of extra compile-connected dependencies. But that's ok.

Note: the protocol causes the ElixirLS server to give some unknown functions warnings. I was unable to disable them (neither with a @dialyzer tag nor with a dialyzer file). Only thing that helped is to disable dialyzer in ElixirLS (or live with the warnings in your editor)