First version with a simple supervision tree
This is a first version with a supervision tree. The Manager should in general be started with Fledex.Supervisor.start_link (except if you have a reason to not bring up the whole system.
All tests are passing, but it's still very rough and a lot of cleanup is required